# core-lending-loan-origination-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-loan-origination
  description: Loan Origination Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/loan-applications/{applicationId}:
    get:
      tags:
        - LoanApplications
      summary: Get loan application by ID
      description: Retrieves a specific loan application.
      operationId: getLoanApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanApplicationDTO'
    put:
      tags:
        - LoanApplications
      summary: Update loan application
      description: Updates an existing loan application record.
      operationId: updateLoanApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplicationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanApplicationDTO'
    delete:
      tags:
        - LoanApplications
      summary: Delete loan application
      description: Deletes a loan application by ID.
      operationId: deleteLoanApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/loan-applications:
    get:
      tags:
        - LoanApplications
      summary: Filter loan applications
      description: Returns a paginated list of loan applications using filtering.
      operationId: filterLoanApplications
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.applicationNumber
          in: query
          description: Exact filter for application number
          required: false
          schema:
            type: string
        - name: filters.applicationStatusId
          in: query
          description: Exact filter for application status id
          required: false
          schema:
            type: string
        - name: filters.applicationSubStatusId
          in: query
          description: Exact filter for application sub status id
          required: false
          schema:
            type: string
        - name: filters.applicationDate
          in: query
          description: Exact filter for application date
          required: false
          schema:
            type: string
        - name: filters.submissionChannelId
          in: query
          description: Exact filter for submission channel id
          required: false
          schema:
            type: string
        - name: filters.distributorId
          in: query
          description: Exact filter for distributor id
          required: false
          schema:
            type: string
        - name: filters.distributorAgencyId
          in: query
          description: Exact filter for distributor agency id
          required: false
          schema:
            type: string
        - name: filters.distributorAgentId
          in: query
          description: Exact filter for distributor agent id
          required: false
          schema:
            type: string
        - name: filters.businessUnit
          in: query
          description: Exact filter for business unit
          required: false
          schema:
            type: string
        - name: filters.branchId
          in: query
          description: Exact filter for branch id
          required: false
          schema:
            type: string
        - name: filters.loanPurpose
          in: query
          description: Exact filter for loan purpose
          required: false
          schema:
            type: string
        - name: filters.loanPurposeDescription
          in: query
          description: Exact filter for loan purpose description
          required: false
          schema:
            type: string
        - name: filters.submittedAt
          in: query
          description: Exact filter for submitted at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[submittedAt].from
          in: query
          description: Filter submitted at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[submittedAt].to
          in: query
          description: Filter submitted at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.firstReviewAt
          in: query
          description: Exact filter for first review at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[firstReviewAt].from
          in: query
          description: Filter first review at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[firstReviewAt].to
          in: query
          description: Filter first review at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.decisionDueDate
          in: query
          description: Exact filter for decision due date
          required: false
          schema:
            type: string
        - name: filters.expectedClosingDate
          in: query
          description: Exact filter for expected closing date
          required: false
          schema:
            type: string
        - name: filters.loanOfficerId
          in: query
          description: Exact filter for loan officer id
          required: false
          schema:
            type: string
        - name: filters.assignedAt
          in: query
          description: Exact filter for assigned at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[assignedAt].from
          in: query
          description: Filter assigned at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[assignedAt].to
          in: query
          description: Filter assigned at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.isRush
          in: query
          description: Exact filter for is rush
          required: false
          schema:
            type: boolean
        - name: filters.isException
          in: query
          description: Exact filter for is exception
          required: false
          schema:
            type: boolean
        - name: filters.requiresManualReview
          in: query
          description: Exact filter for requires manual review
          required: false
          schema:
            type: boolean
        - name: filters.externalReferenceNumber
          in: query
          description: Exact filter for external reference number
          required: false
          schema:
            type: string
        - name: filters.sourceSystem
          in: query
          description: Exact filter for source system
          required: false
          schema:
            type: string
        - name: filters.disbursementMethodType
          in: query
          description: Exact filter for disbursement method type
          required: false
          schema:
            type: string
            enum:
              - INTERNAL_ACCOUNT
              - EXTERNAL_ACCOUNT
        - name: filters.disbursementInternalAccountId
          in: query
          description: Exact filter for disbursement internal account id
          required: false
          schema:
            type: string
        - name: filters.disbursementExternalBankAccountId
          in: query
          description: Exact filter for disbursement external bank account id
          required: false
          schema:
            type: string
        - name: filters.repaymentMethodType
          in: query
          description: Exact filter for repayment method type
          required: false
          schema:
            type: string
            enum:
              - INTERNAL_ACCOUNT
              - EXTERNAL_ACCOUNT
        - name: filters.repaymentInternalAccountId
          in: query
          description: Exact filter for repayment internal account id
          required: false
          schema:
            type: string
        - name: filters.repaymentExternalBankAccountId
          in: query
          description: Exact filter for repayment external bank account id
          required: false
          schema:
            type: string
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseLoanApplicationDTO'
    post:
      tags:
        - LoanApplications
      summary: Create a new loan application
      description: Creates a new loan application record.
      operationId: createLoanApplication
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplicationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanApplicationDTO'
...
  /api/v1/loan-applications/all:
    get:
      tags:
        - LoanApplications
      summary: List all loan applications
      description: Returns a paginated list of loan applications without specific filters.
      operationId: listLoanApplications
      parameters:
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseLoanApplicationDTO'
...
components:
  schemas:
    LoanApplicationDTO:
      required:
        - applicationDate
        - applicationStatusId
        - applicationSubStatusId
        - submissionChannelId
      type: object
      properties:
        loanApplicationId:
          type: string
          format: uuid
          readOnly: true
        applicationNumber:
          type: string
          format: uuid
        applicationStatusId:
          type: string
          format: uuid
        applicationSubStatusId:
          type: string
          format: uuid
        applicationDate:
          type: string
          format: date
        submissionChannelId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        distributorAgencyId:
          type: string
          format: uuid
        distributorAgentId:
          type: string
          format: uuid
        businessUnit:
          type: string
        branchId:
          type: string
          format: uuid
        loanPurpose:
          type: string
        loanPurposeDescription:
          type: string
        submittedAt:
          type: string
          format: date-time
        firstReviewAt:
          type: string
          format: date-time
        decisionDueDate:
          type: string
          format: date
        expectedClosingDate:
          type: string
          format: date
        loanOfficerId:
          type: string
          format: uuid
        assignedAt:
          type: string
          format: date-time
        isRush:
          type: boolean
        isException:
          type: boolean
        requiresManualReview:
          type: boolean
        externalReferenceNumber:
          type: string
        sourceSystem:
          type: string
        disbursementMethodType:
          type: string
          enum:
            - INTERNAL_ACCOUNT
            - EXTERNAL_ACCOUNT
        disbursementInternalAccountId:
          type: string
          format: uuid
        disbursementExternalBankAccountId:
          type: string
          format: uuid
        repaymentMethodType:
          type: string
          enum:
            - INTERNAL_ACCOUNT
            - EXTERNAL_ACCOUNT
        repaymentInternalAccountId:
          type: string
          format: uuid
        repaymentExternalBankAccountId:
          type: string
          format: uuid
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseLoanApplicationDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/LoanApplicationDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...